third_party/blink/web_tests/platform/generic/media/media-fragments/TC0059-expected.txt


Title: TC0059
Fragment: 't=smpte:00:00:01&t=npt:3'
Comment: When a fragment dimensions occurs multiple times, only the last occurrence of that dimension is interpreted.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
